requesting/pyproject.toml
[tool.poetry]
name = "delivering"
version = "0.1.0"
description = ""
authors = ["shin5ok <shin5ok@55mp.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.1"
uvicorn = "^0.22.0"
google-cloud-firestore = "^2.11.1"
Jinja2 = "^3.1.2"
google-cloud-pubsub = "^2.17.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
